requirements.txt
flask==2.3.3
Werkzeug==2.3.7
boto3==1.28.44
python-dotenv==1.0.0
bcrypt==4.0.1
flask-cors==4.0.0
Flask-Session==0.5.0 